# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API para consulta de animes",
        "contact": {
            "name": "Arthur",
            "email": "arthurbrasa@outlook.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1"
    },
    "paths": {
        "/api/v1/anime": {
            "get": {
                "tags": [
                    "Animes"
                ],
                "summary": "Lista de Animes",
                "operationId": "e41e4460d457ed338b894eef6cff2bdd",
                "responses": {
                    "200": {
                        "description": "Lista de Animes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Animes"
                ],
                "summary": "Criar um novo anime",
                "operationId": "93fecc2c1cc26e34a830da70f38275ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Título do anime (máx. 100 caracteres)",
                                        "type": "string"
                                    },
                                    "synopsis": {
                                        "description": "Sinopse do anime",
                                        "type": "string"
                                    },
                                    "release_date": {
                                        "description": "Data de lançamento do anime no formato YYYY-MM-DD",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "image_url": {
                                        "description": "URL da imagem do anime (opcional, máx. 300 caracteres)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Anime criado com sucesso"
                    },
                    "400": {
                        "description": "Requisição inválida"
                    },
                    "409": {
                        "description": "Conflito - Já existe um anime com o mesmo título"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/anime/{title}": {
            "get": {
                "tags": [
                    "Animes"
                ],
                "summary": "Obter informações sobre um anime por título",
                "parameters": [
                    {
                        "name": "title",
                        "in": "path",
                        "description": "Título do anime desejado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações sobre o anime"
                    },
                    "404": {
                        "description": "Anime não encontrado"
                    }
                }
            }
        },
        "/api/v1/anime/{id}": {
            "put": {
                "tags": [
                    "Animes"
                ],
                "summary": "Atualiza um anime existente",
                "operationId": "3e45dd157f722dfa6c04e6be22c10ec2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do anime a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do anime a ser atualizado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Título do anime",
                                        "type": "string"
                                    },
                                    "synopsis": {
                                        "description": "Sinopse do anime",
                                        "type": "string"
                                    },
                                    "release_date": {
                                        "description": "Data de lançamento no formato YYYY-MM-DD",
                                        "type": "date"
                                    },
                                    "image_url": {
                                        "description": "URL da imagem do anime",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Anime atualizado com sucesso"
                    },
                    "404": {
                        "description": "Anime não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Mensagem de erro",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Animes"
                ],
                "summary": "Remove um anime pelo ID",
                "operationId": "c061a276ce7031c88ed9c4408c786cdb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do anime a ser removido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Anime removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensagem de sucesso",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Dados do anime removido",
                                            "type": "object"
                                        },
                                        "status": {
                                            "description": "Código de status da resposta",
                                            "type": "string"
                                        },
                                        "links": {
                                            "description": "Links relacionados",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Anime não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Mensagem de erro",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Mensagem de erro",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}